control-plane/pyproject.toml
[project]
name = "rca-control-plane"
version = "0.0.1"
description = "Remote Control Agent — control plane"
requires-python = ">=3.11"
readme = "../README.md"
license = { text = "Apache-2.0" }

dependencies = [
    # Web/API
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6",
    "python-multipart>=0.0.9",
    # YAML / templating
    "pyyaml>=6.0",
    "jinja2>=3.1",
    # DB
    "sqlalchemy>=2.0",
    "aiosqlite>=0.19",
    # WS / HTTP client
    "websockets>=12",
    "httpx>=0.27",
    # MCP
    # (versione concreta da scegliere nell'ambiente di Fase 1)
    "mcp>=0.1",
    # CLI / util
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.8",
    "httpx>=0.27",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
